+++ package.json
@@ -44,6 +44,7 @@
     "webpack-dev-server": "beta"
   },
   "packages": [
+    "accordion",
     "alert",
     "badge",
     "breadcrumb",
