+++ packages/accordion/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <title>Blaze WC elements catalogue</title>
+</head>
+<body>
+  <bl-<%= htmlWebpackPlugin.options.packages %>-demo>
+</body>
+</html>
