+++ packages/accordion/index.ts
@@ -1,0 +1,2 @@
+import './Accordion';
+import './Accordion.demo';
