+++ packages/collapsible/Collapsible.demo.tsx
@@ -1,6 +1,5 @@
 import { h, Component } from 'skatejs';
 import { Collapsible } from './Collapsible';
-import { CollapsibleItem } from './CollapsibleItem';
 import styles from './Collapsible.scss';
 
 export class Demo extends Component<void> {
@@ -12,31 +11,12 @@ export class Demo extends Component<void> {
         <fieldset>
           <legend>{Collapsible.is}</legend>
 
-          <h3>Single item</h3>
           <Collapsible isOpened>
             <span slot="title">Some <strong>Name</strong></span>
             <div>
               Some content <strong>!!!</strong>
             </div>
           </Collapsible>
-
-          <hr />
-
-          <h3>Two items in card</h3>
-          <div className="c-card">
-            <CollapsibleItem>
-              <span slot="title">Some <strong>Name</strong></span>
-              <div>
-                Some content <strong>!!!</strong>
-              </div>
-            </CollapsibleItem>
-            <CollapsibleItem isOpened isLast>
-              <span slot="title">Some <strong>Name</strong></span>
-              <div>
-                Some content <strong>!!!</strong>
-              </div>
-            </CollapsibleItem>
-          </div>
         </fieldset>
     ]
   }
