+++ packages/collapsible/index.ts
@@ -1,3 +1,2 @@
 import './Collapsible';
-import './CollapsibleItem';
 import './Collapsible.demo';
